fix: DTO refactor
